+++ app/Auth/Application/UseCases/Profile/ProfileOutput.php
@@ -14,8 +14,8 @@ class ProfileOutput extends Output
      */
     public function __construct(
         public readonly string $id,
-        public readonly string $firstName,
-        public readonly string $lastName,
+        public readonly ?string $firstName,
+        public readonly ?string $lastName,
         public readonly string $email,
         public readonly ?string $phoneNumber,
         public readonly ?string $description,