+++ app/Auth/Application/UseCases/UpdateProfile/UpdateProfileInput.php
@@ -10,8 +10,8 @@ class UpdateProfileInput extends Input
 {
     public function __construct(
         public readonly string $userId,
-        public readonly string $firstName,
-        public readonly string $lastName,
+        public readonly ?string $firstName,
+        public readonly ?string $lastName,
         public readonly string $email,
         public readonly ?string $phoneNumber,
         public readonly ?string $description,